Create app that uses child process output events.
Create a node.js app that watches for changes in a file and outputs
information about that file.

Internally, this application watches for changes in a file. When the
file changes, it spawns an child process. The node app reacts to events
emitted by the child process standard output. When the child process
writes a chunk of data, the node application collects that chunk.

When the child process terminates, the node application parses the
output to identify the file permissions, the file size and the file
name. Finally, it prints that information out.